Merge

+++ src/header/string_array_functions.h
@@ -96,9 +96,9 @@ int isAlphanumeric (char *str, int bonus);
  * Retrieve char index in a string
  * @param str the string
  * @param c the char we are looking for
- * @return an index >= 0 or NULL if not found
+ * @return an index >= 0 or -1 if not found
  */
-size_t getCharIndex (char *str, char c);
+int getCharIndex (char *str, char c);
 
 /**
  * Substring a string
@@ -107,7 +107,7 @@ size_t getCharIndex (char *str, char c);
  * @param start index where to start
  * @param length numbers of chars to substring
  */
-void substring (char *str, char *substr, int start, size_t length);
+void substring (char *str, char *substr, int start, int length);
 
 /**
  * Concat two strings into one
